backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
opencv-python==4.10.0.82
numpy==1.26.4
qdrant-client==1.8.2
python-dotenv==1.0.1
pydantic==2.7.1